[WCNCR00262566] wlan: refactor pointer casting
[Description]
refactor uintptr_t as unsigned long for different OS pointer type on x64/32
add define of uintptr_t if not in other OS

MTK-Commit-Id: da16bd5f8f9e49e388ed58b3805a97c3eeb4e398

CR-Id: WCNCR00262566
Feature: os/wlan
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6073805
Tested-by: yuchi.wang <yuchi.wang@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_mspautosanity
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>

+++ include/mgmt/p2p_role.h
@@ -437,11 +437,11 @@ void p2pRoleFsmRunEventCsaDone(IN struct ADAPTER *prAdapter,
 		IN struct MSG_HDR *prMsgHdr);
 
 void p2pRoleFsmRunEventDfsShutDownTimeout(IN struct ADAPTER *prAdapter,
-		IN unsigned long ulParamPtr);
+		IN uintptr_t ulParamPtr);
 #endif
 
 void p2pRoleFsmRunEventWaitNextReqChnlTimeout(IN struct ADAPTER *prAdapter,
-		IN unsigned long ulParamPtr);
+		IN uintptr_t ulParamPtr);
 
 void p2pRoleFsmRunEventScanRequest(IN struct ADAPTER *prAdapter,
 		IN struct MSG_HDR *prMsgHdr);
@@ -455,10 +455,10 @@ void p2pRoleFsmRunEventJoinComplete(IN struct ADAPTER *prAdapter,
 		IN struct MSG_HDR *prMsgHdr);
 
 void p2pRoleFsmRunEventTimeout(IN struct ADAPTER *prAdapter,
-		IN unsigned long ulParamPtr);
+		IN uintptr_t ulParamPtr);
 
 void p2pRoleFsmDeauthTimeout(IN struct ADAPTER *prAdapter,
-		IN unsigned long ulParamPtr);
+		IN uintptr_t ulParamPtr);
 
 void p2pRoleFsmRunEventBeaconTimeout(IN struct ADAPTER *prAdapter,
 		IN struct BSS_INFO *prP2pBssInfo);
@@ -536,7 +536,7 @@ p2pProcessEvent_UpdateNOAParam(IN struct ADAPTER *prAdapter,
 
 #if	CFG_ENABLE_PER_STA_STATISTICS_LOG
 void p2pRoleFsmGetStaStatistics(IN struct ADAPTER *prAdapter,
-		IN unsigned long ulParamPtr);
+		IN uintptr_t ulParamPtr);
 #endif
 
 void p2pRoleFsmNotifyEapolTxStatus(IN struct ADAPTER *prAdapter,
